check, deposit and withdraw balance routes done

+++ src/api/migrations/20230322091815-create-orders.js
@@ -22,7 +22,7 @@ module.exports = {
 				type: Sequelize.INTEGER,
 			},
 			totalPrice: {
-				type: Sequelize.INTEGER,
+				type: Sequelize.FLOAT(10, 2),
 			},
 			createdAt: {
 				allowNull: false,
